Delete file and combine modes draft
