+++ CPK-PYTRIS/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/Users/chuyeee/opt/anaconda3/bin/python"
+}
