+++ CPK-PYTRIS/leaderboard.txt
@@ -1,3 +1,12 @@
 BRB 1290
 IJE 27410
 BYB 1270
+AAA 110
+AAA 110
+AAA 110
+AAA 110
+AAA 110
+CPK 110
+AAA 120
+CAA 120
+AAA 110
