--- CPK-PYTRIS/mino.old.py
@@ -1,190 +1,0 @@
-class tetrimino:
-########################################### I
-    I = [
-        [
-            [0, 0, 0, 0],
-            [1, 1, 1, 1],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 1, 0],
-            [0, 0, 1, 0],
-            [0, 0, 1, 0],
-            [0, 0, 1, 0]
-        ],
-        [
-            [0, 0, 0, 0],
-            [0, 0, 0, 0],
-            [1, 1, 1, 1],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 1, 0, 0],
-            [0, 1, 0, 0],
-            [0, 1, 0, 0],
-            [0, 1, 0, 0]
-        ]
-    ]
-########################################### J
-    J = [
-        [
-            [2, 0, 0, 0],
-            [2, 2, 2, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 2, 2, 0],
-            [0, 2, 0, 0],
-            [0, 2, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 0, 0],
-            [2, 2, 2, 0],
-            [0, 0, 2, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 2, 0, 0],
-            [0, 2, 0, 0],
-            [2, 2, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    ]
-########################################### L
-    L = [
-        [
-            [0, 0, 3, 0],
-            [3, 3, 3, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 3, 0, 0],
-            [0, 3, 0, 0],
-            [0, 3, 3, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 0, 0],
-            [3, 3, 3, 0],
-            [3, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [3, 3, 0, 0],
-            [0, 3, 0, 0],
-            [0, 3, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    ]
-########################################### O
-    O = [
-        [
-            [0, 4, 4, 0],
-            [0, 4, 4, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 4, 4, 0],
-            [0, 4, 4, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 4, 4, 0],
-            [0, 4, 4, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 4, 4, 0],
-            [0, 4, 4, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    ]
-########################################### S
-    S = [
-        [
-            [0, 5, 5, 0],
-            [5, 5, 0, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 5, 0, 0],
-            [0, 5, 5, 0],
-            [0, 0, 5, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 0, 0],
-            [0, 5, 5, 0],
-            [5, 5, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [5, 0, 0, 0],
-            [5, 5, 0, 0],
-            [0, 5, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    ]
-########################################### T
-    T = [
-        [
-            [0, 6, 0, 0],
-            [6, 6, 6, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 6, 0, 0],
-            [0, 6, 6, 0],
-            [0, 6, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 0, 0],
-            [6, 6, 6, 0],
-            [0, 6, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 6, 0, 0],
-            [6, 6, 0, 0],
-            [0, 6, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    ]
-########################################### Z
-    Z = [
-        [
-            [7, 7, 0, 0],
-            [0, 7, 7, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 7, 0],
-            [0, 7, 7, 0],
-            [0, 7, 0, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 0, 0, 0],
-            [7, 7, 0, 0],
-            [0, 7, 7, 0],
-            [0, 0, 0, 0]
-        ],
-        [
-            [0, 7, 0, 0],
-            [7, 7, 0, 0],
-            [7, 0, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    ]
