--- CPK-PYTRIS/run.py
@@ -1,771 +1,0 @@
-# PYTRIS™ Copyright (c) 2017 Jason Kim All Rights Reserved.
-
-import pygame
-import operator
-from mino import *
-from random import *
-from pygame.locals import *
-
-# Define
-block_size = 17  # Height, width of single block
-width = 10  # Board width
-height = 20  # Board height
-framerate = 30
-
-board_width = 750
-board_height = 600
-
-pygame.init()
-
-clock = pygame.time.Clock() #FPS
-screen = pygame.display.set_mode((board_width, board_height)) #창크기
-pygame.time.set_timer(pygame.USEREVENT, framerate * 10)
-pygame.display.set_caption("CPKTRIS!™")
-
-class ui_variables:
-    # Fonts
-    font_path = "./assets/fonts/DungGeunMo.ttf"
-    h1 = pygame.font.Font(font_path, 50)
-    h2 = pygame.font.Font(font_path, 30)
-    h4 = pygame.font.Font(font_path, 20)
-    h5 = pygame.font.Font(font_path, 13)
-    h6 = pygame.font.Font(font_path, 10)
-
-    # Sounds
-    click_sound = pygame.mixer.Sound("assets/sounds/SFX_ButtonUp.wav")
-    move_sound = pygame.mixer.Sound("assets/sounds/SFX_PieceMoveLR.wav")
-    drop_sound = pygame.mixer.Sound("assets/sounds/SFX_PieceHardDrop.wav")
-    single_sound = pygame.mixer.Sound("assets/sounds/SFX_SpecialLineClearSingle.wav")
-    double_sound = pygame.mixer.Sound("assets/sounds/SFX_SpecialLineClearDouble.wav")
-    triple_sound = pygame.mixer.Sound("assets/sounds/SFX_SpecialLineClearTriple.wav")
-    tetris_sound = pygame.mixer.Sound("assets/sounds/SFX_SpecialTetris.wav")
-
-    # Background colors
-    black = (10, 10, 10)  # rgb(10, 10, 10)
-    white = (255, 255, 255)  # rgb(255, 255, 255)
-    grey_1 = (26, 26, 26)  # rgb(26, 26, 26)
-    grey_2 = (35, 35, 35)  # rgb(35, 35, 35)
-    grey_3 = (55, 55, 55)  # rgb(55, 55, 55)
-
-    # Tetrimino colors
-    cyan = (69, 206, 204)  # rgb(69, 206, 204) # I
-    blue = (64, 111, 249)  # rgb(64, 111, 249) # J
-    orange = (253, 189, 53)  # rgb(253, 189, 53) # L
-    yellow = (246, 227, 90)  # rgb(246, 227, 90) # O
-    green = (98, 190, 68)  # rgb(98, 190, 68) # S
-    pink = (242, 64, 235)  # rgb(242, 64, 235) # T
-    red = (225, 13, 27)  # rgb(225, 13, 27) # Z
-
-    t_color = [grey_2, cyan, blue, orange, yellow, green, pink, red, grey_3]
-
-#image
-background_image = 'assets/images/retrotv.jpg'
-main_image = 'assets/images/main.png'
-main = pygame.image.load(main_image)
-main = pygame.transform.scale(main,  (board_width, board_height))
-
-#objects
-text1 = ui_variables.h2.render("original", 1, ui_variables.white)
-text2 = ui_variables.h2.render("original", 1, ui_variables.red)
-text3 = ui_variables.h2.render("rotate", 1, ui_variables.white)
-text4 = ui_variables.h2.render("rotate", 1, ui_variables.red)
-text5 = ui_variables.h2.render("dual screen", 1, ui_variables.white)
-text6 = ui_variables.h2.render("dual screen", 1, ui_variables.red)
-text7 = ui_variables.h2.render("blackout", 1, ui_variables.white)
-text8 = ui_variables.h2.render("blackout", 1, ui_variables.red)
-text9 = ui_variables.h2.render("information", 1, ui_variables.white)
-text10 = ui_variables.h2.render("information", 1, ui_variables.red)
-
-class button(): #버튼객체
-    def __init__(self, board_width, board_height, x_rate, y_rate, width_rate, height_rate, text): #버튼생성
-        self.x = board_width * x_rate #버튼 x좌표
-        self.y = board_height * y_rate #버튼 y좌표
-        self.width = int(board_width * width_rate) #버튼 너비
-        self.height = int(board_height * height_rate) #버튼 높이
-        self.x_rate = x_rate #board_width * x_rate = x좌표
-        self.y_rate = y_rate #board_height * y_rate = y좌표
-        self.width_rate = width_rate #board_width * width_rate = 버튼 너비
-        self.height_rate = height_rate #board_height * height_rate = 버튼 높이
-        self.text = text #라벨
-
-    def change(self, board_width, board_height): #버튼 위치, 크기 바꾸기
-        self.x = board_width * self.x_rate #x좌표
-        self.y = board_height * self.y_rate #y좌표
-        self.width = int(board_width * self.width_rate) #너비
-        self.height = int(board_height * self.height_rate) #높이
-
-    def draw(self, win, outline=None): #버튼 보이게 만들기
-        if outline:
-            draw_text(screen, self.text, self.x, self.y, self.width, self.height)
-
-    def isOver(self, pos): #마우스의 위치에 따라 버튼 누르기 pos[0]은 마우스 x좌표, pos[1]은 마우스 y좌표
-        if pos[0] > self.x - (self.width / 2) and pos[0] < self.x + (self.width / 2):
-            if pos[1] > self.y - (self.height / 2) and pos[1] < self.y + (self.height / 2):
-                return True
-        return False
-
-    def isOver_2(self, pos): #start 화면에서 single,pvp,help,setting을 위해서 y좌표 좁게 인식하도록
-        if pos[0] > self.x - (self.width / 2) and pos[0] < self.x + (self.width / 2):
-            if pos[1] > self.y - (self.height / 8) and pos[1] < self.y + (self.height / 8):#243줄에서의 2을 4로 바꿔주면서 좁게 인식할수 있도록함. 더 좁게 인식하고 싶으면 숫자 늘려주기#
-                return True
-        return False
-
-#button
-origianl_bnt = button(board_width, board_height, 0.32, 0.35, 0.2, 0.4, text1)
-rotate_bnt = button(board_width, board_height, 0.32, 0.45, 0.2, 0.4, text3)
-dual_bnt = button(board_width, board_height, 0.32, 0.55, 0.2, 0.4, text5)
-blackout_bnt = button(board_width, board_height, 0.32, 0.65, 0.2, 0.4, text7)
-info_bnt = button(board_width, board_height, 0.32, 0.75, 0.2, 0.4, text9)
-
-bnt_list = [origianl_bnt, blackout_bnt, rotate_bnt, info_bnt, dual_bnt]
-
-#Draw button text
-def draw_text(window, text, x, y, width, height):
-    x = x - (width / 2)
-    window.blit(text, (x, y))
-
-# Draw block
-def draw_block(x, y, color):
-    pygame.draw.rect(screen, color, Rect(x, y, block_size, block_size))
-    pygame.draw.rect(screen, ui_variables.grey_1, Rect(x, y, block_size, block_size), 1)
-
-
-# Draw game screen
-def draw_board(next, hold, score, level, goal):
-    screen.fill(ui_variables.cyan)
-
-    # Draw sidebar
-    pygame.draw.rect(screen, ui_variables.white, Rect(204 + 188, 0 + 113, 96 + 74, 374))
-
-    # Draw next mino
-    grid_n = tetrimino.mino_map[next - 1][0]
-
-    for i in range(4):
-        for j in range(4):
-            dx = 220 + 188 + block_size * j
-            dy = 140 + 113 + block_size * i
-            if grid_n[i][j] != 0:
-                pygame.draw.rect(
-                    screen,
-                    ui_variables.t_color[grid_n[i][j]],
-                    Rect(dx, dy, block_size, block_size),
-                )
-
-    # Draw hold mino
-    grid_h = tetrimino.mino_map[hold - 1][0]
-
-    if hold_mino != -1:
-        for i in range(4):
-            for j in range(4):
-                dx = 220 + 188 + block_size * j
-                dy = 50 + 113 + block_size * i
-                if grid_h[i][j] != 0:
-                    pygame.draw.rect(
-                        screen,
-                        ui_variables.t_color[grid_h[i][j]],
-                        Rect(dx, dy, block_size, block_size),
-                    )
-
-    # Set max score
-    if score > 999999:
-        score = 999999
-
-    # Draw texts
-    text_hold = ui_variables.h5.render("HOLD", 1, ui_variables.black)
-    text_next = ui_variables.h5.render("NEXT", 1, ui_variables.black)
-    text_score = ui_variables.h5.render("SCORE", 1, ui_variables.black)
-    score_value = ui_variables.h4.render(str(score), 1, ui_variables.black)
-    text_level = ui_variables.h5.render("LEVEL", 1, ui_variables.black)
-    level_value = ui_variables.h4.render(str(level), 1, ui_variables.black)
-    text_goal = ui_variables.h5.render("GOAL", 1, ui_variables.black)
-    goal_value = ui_variables.h4.render(str(goal), 1, ui_variables.black)
-
-    # Place texts
-    screen.blit(text_hold, (215 + 188, 14 + 113))
-    screen.blit(text_next, (215 + 188, 104 + 113))
-    screen.blit(text_score, (215 + 188, 194 + 113))
-    screen.blit(score_value, (220 + 188, 210 + 113))
-    screen.blit(text_level, (215 + 188, 254 + 113))
-    screen.blit(level_value, (220 + 188, 270 + 113))
-    screen.blit(text_goal, (215 + 188, 314 + 113))
-    screen.blit(goal_value, (220 + 188, 330 + 113))
-
-    # Draw board
-    for x in range(width):
-        for y in range(height):
-            dx = 17 + 188 + block_size * x
-            dy = 17 + 113 + block_size * y
-            draw_block(dx, dy, ui_variables.t_color[matrix[x][y + 1]])
-
-
-# Draw a tetrimino
-def draw_mino(x, y, mino, r):
-    grid = tetrimino.mino_map[mino - 1][r]
-
-    tx, ty = x, y
-    while not is_bottom(tx, ty, mino, r):
-        ty += 1
-
-    # Draw ghost
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                matrix[tx + j][ty + i] = 8
-
-    # Draw mino
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                matrix[x + j][y + i] = grid[i][j]
-
-
-# Erase a tetrimino
-def erase_mino(x, y, mino, r):
-    grid = tetrimino.mino_map[mino - 1][r]
-
-    # Erase ghost
-    for j in range(21):
-        for i in range(10):
-            if matrix[i][j] == 8:
-                matrix[i][j] = 0
-
-    # Erase mino
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                matrix[x + j][y + i] = 0
-
-
-# Returns true if mino is at bottom
-def is_bottom(x, y, mino, r):
-    grid = tetrimino.mino_map[mino - 1][r]
-
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                if (y + i + 1) > 20:
-                    return True
-                elif matrix[x + j][y + i + 1] != 0 and matrix[x + j][y + i + 1] != 8:
-                    return True
-
-    return False
-
-
-# Returns true if mino is at the left edge
-def is_leftedge(x, y, mino, r):
-    grid = tetrimino.mino_map[mino - 1][r]
-
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                if (x + j - 1) < 0:
-                    return True
-                elif matrix[x + j - 1][y + i] != 0:
-                    return True
-
-    return False
-
-
-# Returns true if mino is at the right edge
-def is_rightedge(x, y, mino, r):
-    grid = tetrimino.mino_map[mino - 1][r]
-
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                if (x + j + 1) > 9:
-                    return True
-                elif matrix[x + j + 1][y + i] != 0:
-                    return True
-
-    return False
-
-
-# Returns true if turning right is possible
-def is_turnable_r(x, y, mino, r):
-    if r != 3:
-        grid = tetrimino.mino_map[mino - 1][r + 1]
-    else:
-        grid = tetrimino.mino_map[mino - 1][0]
-
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                if (x + j) < 0 or (x + j) > 9 or (y + i) < 0 or (y + i) > 20:
-                    return False
-                elif matrix[x + j][y + i] != 0:
-                    return False
-
-    return True
-
-
-# Returns true if turning left is possible
-def is_turnable_l(x, y, mino, r):
-    if r != 0:
-        grid = tetrimino.mino_map[mino - 1][r - 1]
-    else:
-        grid = tetrimino.mino_map[mino - 1][3]
-
-    for i in range(4):
-        for j in range(4):
-            if grid[i][j] != 0:
-                if (x + j) < 0 or (x + j) > 9 or (y + i) < 0 or (y + i) > 20:
-                    return False
-                elif matrix[x + j][y + i] != 0:
-                    return False
-
-    return True
-
-
-# Returns true if new block is drawable
-def is_stackable(mino):
-    grid = tetrimino.mino_map[mino - 1][0]
-
-    for i in range(4):
-        for j in range(4):
-            # print(grid[i][j], matrix[3 + j][i])
-            if grid[i][j] != 0 and matrix[3 + j][i] != 0:
-                return False
-
-    return True
-
-
-# Initial values
-blink = False
-start = False
-pause = False
-done = False
-game_over = False
-original = False
-blackout = False
-dual = False
-info = False
-
-score = 0
-level = 1
-goal = level * 5
-bottom_count = 0
-hard_drop = False
-
-dx, dy = 3, 0  # Minos location status
-rotation = 0  # Minos rotation status
-
-mino = randint(1, 7)  # Current mino
-next_mino = randint(1, 7)  # Next mino
-
-hold = False  # Hold status
-hold_mino = -1  # Holded mino
-
-name_location = 0
-name = [65, 65, 65]
-
-with open("leaderboard.txt") as f:
-    lines = f.readlines()
-lines = [line.rstrip("\n") for line in open("leaderboard.txt")]
-
-leaders = {"AAA": 0, "BBB": 0, "CCC": 0}
-for i in lines:
-    leaders[i.split(" ")[0]] = int(i.split(" ")[1])
-leaders = sorted(leaders.items(), key=operator.itemgetter(1), reverse=True)
-
-matrix = [[0 for y in range(height + 1)] for x in range(width)]  # Board matrix
-
-###########################################################
-# Loop Start
-###########################################################
-
-while not done:
-    # Pause screen
-    if pause:
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                done = True
-            elif event.type == USEREVENT:
-                pygame.time.set_timer(pygame.USEREVENT, 300)
-                draw_board(next_mino, hold_mino, score, level, goal)
-
-                pause_text = ui_variables.h2_b.render("PAUSED", 1, ui_variables.white)
-                pause_start = ui_variables.h5.render(
-                    "Press esc to continue", 1, ui_variables.white
-                )
-
-                screen.blit(pause_text, (43, 100))
-                if blink:
-                    screen.blit(pause_start, (40, 160))
-                    blink = False
-                else:
-                    blink = True
-                pygame.display.update()
-            elif event.type == KEYDOWN:
-                erase_mino(dx, dy, mino, rotation)
-                if event.key == K_ESCAPE:
-                    pause = False
-                    ui_variables.click_sound.play()
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-
-    # Game screen
-    elif start:
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                done = True
-            elif event.type == USEREVENT:
-                # Set speed
-                if not game_over:
-                    keys_pressed = pygame.key.get_pressed()
-                    if keys_pressed[K_DOWN]:
-                        pygame.time.set_timer(pygame.USEREVENT, framerate * 1)
-                    else:
-                        pygame.time.set_timer(pygame.USEREVENT, framerate * 10)
-
-                # Draw a mino
-                draw_mino(dx, dy, mino, rotation)
-                draw_board(next_mino, hold_mino, score, level, goal)
-
-                # Erase a mino
-                if not game_over:
-                    erase_mino(dx, dy, mino, rotation)
-
-                # Move mino down
-                if not is_bottom(dx, dy, mino, rotation):
-                    dy += 1
-
-                # Create new mino
-                else:
-                    if hard_drop or bottom_count == 6:
-                        hard_drop = False
-                        bottom_count = 0
-                        score += 10 * level
-                        draw_mino(dx, dy, mino, rotation)
-                        draw_board(next_mino, hold_mino, score, level, goal)
-                        if is_stackable(next_mino):
-                            mino = next_mino
-                            next_mino = randint(1, 7)
-                            dx, dy = 3, 0
-                            rotation = 0
-                            hold = False
-                        else:
-                            start = False
-                            game_over = True
-                            pygame.time.set_timer(pygame.USEREVENT, 1)
-                    else:
-                        bottom_count += 1
-
-                # Erase line
-                erase_count = 0
-                for j in range(21):
-                    is_full = True
-                    for i in range(10):
-                        if matrix[i][j] == 0:
-                            is_full = False
-                    if is_full:
-                        erase_count += 1
-                        k = j
-                        while k > 0:
-                            for i in range(10):
-                                matrix[i][k] = matrix[i][k - 1]
-                            k -= 1
-                if erase_count == 1:
-                    ui_variables.single_sound.play()
-                    score += 50 * level
-                elif erase_count == 2:
-                    ui_variables.double_sound.play()
-                    score += 150 * level
-                elif erase_count == 3:
-                    ui_variables.triple_sound.play()
-                    score += 350 * level
-                elif erase_count == 4:
-                    ui_variables.tetris_sound.play()
-                    score += 1000 * level
-
-                # Increase level
-                goal -= erase_count
-                if goal < 1 and level < 15:
-                    level += 1
-                    goal += level * 5
-                    framerate = int(framerate * 0.8)
-
-            elif event.type == KEYDOWN:
-                erase_mino(dx, dy, mino, rotation)
-                if event.key == K_ESCAPE:
-                    ui_variables.click_sound.play()
-                    pause = True
-                # Hard drop
-                elif event.key == K_SPACE:
-                    ui_variables.drop_sound.play()
-                    while not is_bottom(dx, dy, mino, rotation):
-                        dy += 1
-                    hard_drop = True
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-                    draw_mino(dx, dy, mino, rotation)
-                    draw_board(next_mino, hold_mino, score, level, goal)
-                # Hold
-                elif event.key == K_LSHIFT or event.key == K_c:
-                    if hold == False:
-                        ui_variables.move_sound.play()
-                        if hold_mino == -1:
-                            hold_mino = mino
-                            mino = next_mino
-                            next_mino = randint(1, 7)
-                        else:
-                            hold_mino, mino = mino, hold_mino
-                        dx, dy = 3, 0
-                        rotation = 0
-                        hold = True
-                    draw_mino(dx, dy, mino, rotation)
-                    draw_board(next_mino, hold_mino, score, level, goal)
-                # Turn right
-                elif event.key == K_UP or event.key == K_x:
-                    if is_turnable_r(dx, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        rotation += 1
-                    # Kick
-                    elif is_turnable_r(dx, dy - 1, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dy -= 1
-                        rotation += 1
-                    elif is_turnable_r(dx + 1, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx += 1
-                        rotation += 1
-                    elif is_turnable_r(dx - 1, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx -= 1
-                        rotation += 1
-                    elif is_turnable_r(dx, dy - 2, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dy -= 2
-                        rotation += 1
-                    elif is_turnable_r(dx + 2, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx += 2
-                        rotation += 1
-                    elif is_turnable_r(dx - 2, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx -= 2
-                        rotation += 1
-                    if rotation == 4:
-                        rotation = 0
-                    draw_mino(dx, dy, mino, rotation)
-                    draw_board(next_mino, hold_mino, score, level, goal)
-                # Turn left
-                elif event.key == K_z or event.key == K_LCTRL:
-                    if is_turnable_l(dx, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        rotation -= 1
-                    # Kick
-                    elif is_turnable_l(dx, dy - 1, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dy -= 1
-                        rotation -= 1
-                    elif is_turnable_l(dx + 1, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx += 1
-                        rotation -= 1
-                    elif is_turnable_l(dx - 1, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx -= 1
-                        rotation -= 1
-                    elif is_turnable_l(dx, dy - 2, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dy -= 2
-                        rotation += 1
-                    elif is_turnable_l(dx + 2, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx += 2
-                        rotation += 1
-                    elif is_turnable_l(dx - 2, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx -= 2
-                    if rotation == -1:
-                        rotation = 3
-                    draw_mino(dx, dy, mino, rotation)
-                    draw_board(next_mino, hold_mino, score, level, goal)
-                # Move left
-                elif event.key == K_LEFT:
-                    if not is_leftedge(dx, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx -= 1
-                    draw_mino(dx, dy, mino, rotation)
-                    draw_board(next_mino, hold_mino, score, level, goal)
-                # Move right
-                elif event.key == K_RIGHT:
-                    if not is_rightedge(dx, dy, mino, rotation):
-                        ui_variables.move_sound.play()
-                        dx += 1
-                    draw_mino(dx, dy, mino, rotation)
-                    draw_board(next_mino, hold_mino, score, level, goal)
-
-        pygame.display.update()
-
-    # Game over screen
-    elif game_over:
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                done = True
-            elif event.type == USEREVENT:
-                pygame.time.set_timer(pygame.USEREVENT, 300)
-                over_text_1 = ui_variables.h2_b.render("GAME", 1, ui_variables.white)
-                over_text_2 = ui_variables.h2_b.render("OVER", 1, ui_variables.white)
-                over_start = ui_variables.h5.render(
-                    "Press return to continue", 1, ui_variables.white
-                )
-
-                draw_board(next_mino, hold_mino, score, level, goal)
-                screen.blit(over_text_1, (58, 75))
-                screen.blit(over_text_2, (62, 105))
-
-                name_1 = ui_variables.h2_i.render(chr(name[0]), 1, ui_variables.white)
-                name_2 = ui_variables.h2_i.render(chr(name[1]), 1, ui_variables.white)
-                name_3 = ui_variables.h2_i.render(chr(name[2]), 1, ui_variables.white)
-
-                underbar_1 = ui_variables.h2.render("_", 1, ui_variables.white)
-                underbar_2 = ui_variables.h2.render("_", 1, ui_variables.white)
-                underbar_3 = ui_variables.h2.render("_", 1, ui_variables.white)
-
-                screen.blit(name_1, (65, 147))
-                screen.blit(name_2, (95, 147))
-                screen.blit(name_3, (125, 147))
-
-                if blink:
-                    screen.blit(over_start, (32, 195))
-                    blink = False
-                else:
-                    if name_location == 0:
-                        screen.blit(underbar_1, (65, 145))
-                    elif name_location == 1:
-                        screen.blit(underbar_2, (95, 145))
-                    elif name_location == 2:
-                        screen.blit(underbar_3, (125, 145))
-                    blink = True
-
-                pygame.display.update()
-            elif event.type == KEYDOWN:
-                if event.key == K_RETURN:
-                    ui_variables.click_sound.play()
-
-                    outfile = open("leaderboard.txt", "a")
-                    outfile.write(
-                        chr(name[0])
-                        + chr(name[1])
-                        + chr(name[2])
-                        + " "
-                        + str(score)
-                        + "\n"
-                    )
-                    outfile.close()
-
-                    game_over = False
-                    hold = False
-                    dx, dy = 3, 0
-                    rotation = 0
-                    mino = randint(1, 7)
-                    next_mino = randint(1, 7)
-                    hold_mino = -1
-                    framerate = 30
-                    score = 0
-                    score = 0
-                    level = 1
-                    goal = level * 5
-                    bottom_count = 0
-                    hard_drop = False
-                    name_location = 0
-                    name = [65, 65, 65]
-                    matrix = [[0 for y in range(height + 1)] for x in range(width)]
-
-                    with open("leaderboard.txt") as f:
-                        lines = f.readlines()
-                    lines = [line.rstrip("\n") for line in open("leaderboard.txt")]
-
-                    leaders = {"AAA": 0, "BBB": 0, "CCC": 0}
-                    for i in lines:
-                        leaders[i.split(" ")[0]] = int(i.split(" ")[1])
-                    leaders = sorted(
-                        leaders.items(), key=operator.itemgetter(1), reverse=True
-                    )
-
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-                elif event.key == K_RIGHT:
-                    if name_location != 2:
-                        name_location += 1
-                    else:
-                        name_location = 0
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-                elif event.key == K_LEFT:
-                    if name_location != 0:
-                        name_location -= 1
-                    else:
-                        name_location = 2
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-                elif event.key == K_UP:
-                    ui_variables.click_sound.play()
-                    if name[name_location] != 90:
-                        name[name_location] += 1
-                    else:
-                        name[name_location] = 65
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-                elif event.key == K_DOWN:
-                    ui_variables.click_sound.play()
-                    if name[name_location] != 65:
-                        name[name_location] -= 1
-                    else:
-                        name[name_location] = 90
-                    pygame.time.set_timer(pygame.USEREVENT, 1)
-
-    elif original:
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                done = True
-            elif event.type == USEREVENT:
-                start = True
-
-
-    # Start Screen
-    else:
-        for event in pygame.event.get():
-            pos = pygame.mouse.get_pos()
-            if event.type == QUIT:
-                done = True
-
-            elif event.type == pygame.MOUSEMOTION:
-                if origianl_bnt.isOver_2(pos):
-                    origianl_bnt.text = text2
-                else:
-                    origianl_bnt.text = text1
-                if rotate_bnt.isOver_2(pos):
-                    rotate_bnt.text = text4
-                else:
-                    rotate_bnt.text = text3
-                if dual_bnt.isOver_2(pos):
-                    dual_bnt.text = text6
-                else:
-                    dual_bnt.text = text5
-                if blackout_bnt.isOver_2(pos):
-                    blackout_bnt.text = text8
-                else:
-                    blackout_bnt.text = text7
-                if info_bnt.isOver_2(pos):
-                    info_bnt.text = text10
-                else:
-                    info_bnt.text = text9
-
-            elif event.type == pygame.MOUSEBUTTONDOWN:
-                if origianl_bnt.isOver_2(pos):
-                    ui_variables.click_sound.play()
-                    original= True
-
-        # pygame.time.set_timer(pygame.USEREVENT, 300)
-        screen.fill(ui_variables.white)
-        screen.blit(main, [0,0])
-
-        origianl_bnt.draw(screen, (0, 0, 0))
-        rotate_bnt.draw(screen, (0, 0, 0))
-        dual_bnt.draw(screen, (0, 0, 0))
-        blackout_bnt.draw(screen, (0, 0, 0))
-        info_bnt.draw(screen, (0, 0, 0))
-
-
-        if not start:
-            pygame.display.update()
-            clock.tick(3)
-
-pygame.quit()
